arch: Tier-1 follow-up — carve dataSource.js loaders into per-layer files
dataSource.js loadManifest() used to be a 250-line if/else if chain
with one branch per layer (sst7d, sst5d, swell5d, wind5d, current5d,
wind, viz, sst, chl). Each branch was 30-50 lines of nearly-identical
"fetch summary → decode PNGs → write state.layers[name]" code.

That shape is exactly what made `no-dupe-else-if` load-bearing for
the 2026-05-07 white-screen incident: my Phase E sst5d loader and
main's parallel sst5d loader collided in the same chain, and ESLint
caught a duplicate `else if (layer === "sst5d")` branch. The audit
called it out as Tier 1 #3.

This refactor:

  src/lib/loaders/
    decoders.js      — shared helpers (decodePng, decodeUVPng,
                       decodeWavePng, fillNearestInPlace,
                       computeSpeedKt, currentSampleMask,
                       landMaskedCurrentSample, bucketKey, hourKey)
    sst7d.js         — loadSst7d(info, state)
    sst5d.js         — loadSst5d(info, state)
    swell5d.js       — loadSwell5d(info, state)
    wind5d.js        — loadWind5d(info, state)
    current5d.js     — loadCurrent5d(info, state)
    wind.js          — loadWind(info, state)
    viz.js           — loadViz(info, state)
    scalarPng.js     — loadScalarPng(layer, info, state) for sst/chl
    index.js         — LAYER_LOADERS registry + decoder re-exports

dataSource.js loadManifest() becomes a 5-line registry dispatch:

    for (const [layer, info] of Object.entries(manifest.layers || {})) {
      state.layers[layer] = layer === "sst" ? (state.layers.sst || {}) : {};
      const loader = LAYER_LOADERS[layer];
      if (loader) await loader(info, state);
    }

Behaviour is byte-for-byte equivalent — every loader function is the
exact body of its old branch, lifted out. The only logical change is
that wave + precip (which were silently unreachable in the old chain)
now route through the registry's "no entry → skip" path. Same effect
either way.

bucketKey / hourKey moved to decoders.js but dataSource.js re-exports
them, so CurrentTimeline.jsx + WindDayGrid.jsx (which import from
dataSource.js) don't need any change.

Net file size:
  dataSource.js  1,185 → 798 LOC (-32%)
  9 new files in src/lib/loaders/  +587 LOC across the layout

The +200 LOC overhead is all import statements + per-file headers.
Each individual loader is now ~30-70 lines, reviewable in one screen,
and a new layer is one new file + one registry entry — not a 50-line
diff in the middle of a god-function.
